tests: run against CentOS 8 versions of dependencies
This is a first step to add RPM building to the project. It adds a new
Nox session which doesn't install dependencies as pinned by Poetry (in
`poetry.lock`), but instead hard-codes some versions/constraints as
found in CentOS 8 and EPEL.

Since the original tests code was not compatible with these older
versions of packages (e.g., `pytest`), some code changes were required,
either based on import names being (un)available, or based on version
checks.