tests: increase coverage on all Python versions
Given this change, both the `os.preadv`-backed implementation of
`gptsum.checksum.hash_file` as well as the `os.pread`-backed version are
covered by the test-suite. This means that, in a sufficiently recent
Python environment (>= 3.7), the code is now almost 100% covered by the
test-suite: only the import of `importlib_metadata` (or
`importlib.metadata`) isn't fully covered: either one or the other is
chosen based on the Python version (the former for Python <= 3.7, the
latter when using Python >= 3.8).