ci: enable conditional coverage markers
Using the `coverage-conditional-plugin`, we can set pragmas on specific
code sections which allow coverage to be skipped based on conditions. As
such, code that won't work on Python 3.6 (using `os.preadv`), or on
Python >= 3.8 (where `importlib_metadata` is no longer used) can be
marked as such, and we regain 100% coverage, for each Python version
individually.

The current version of `coverage-conditional-plugin` is (supposedly) not
compatible with `coverage >=6.0`, hence this dependency is reverted to
the latest 5.x version, for now: in Git, `coverage-conditional-plugin`
did get support for `coverage ^6.0`.

See: https://github.com/wemake-services/coverage-conditional-plugin
See: https://sobolevn.me/2020/02/conditional-coverage