ci: enforce 100% coverage for every Python version
By running `coverage report` as part of the `tests` session, 100%
coverage is now required for every (supported) Python version.

This is achieved by letting the coverage data be stored in a specific
file rather than a random one.

The aggregated coverage data is still combined and validated for
`codecov` to function.