ci: always run `tests-success` stage
When the `tests-success` stage is skipped, it's not counted as a
prerequisite/status-check to pass before merging is allowed. At the same
time, when a job `needs` others, and its dependencies fail, it's
skipped, unless it's configured to always run.

Hence, we need this.

See: https://docs.github.com/en/actions/learn-github-actions/workflow-syntax-for-github-actions#jobsjob_idneeds