tests: `close` a `NamedTemporaryFile` before passing it on
A `NamedTemporaryFile` includes a file-descriptor to the file, which we
don't really use (the file path is passed on instead). However, on
Windows, this is an issue: if test code laters attempts to open the
file, this fails with a `PermissionError` (can't open a file twice).

Hence, close the file-descriptor before yielding the filename to the
fixture user, but also make sure the call to `close` won't also delete
the file.